# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for MyBank Application",
        "title": "MyBank API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of transactions for the logged-in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.TransactionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new transaction (transfer, payment, topup, withdraw, deposit, qris)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create a new transaction",
                "parameters": [
                    {
                        "description": "Create Transaction Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_request.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.TransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    }
                }
            }
        },
        "/transactions/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a summary of user spending including total spend, favourite category, weekly comparisons, and top merchants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get spending summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.SpendingSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a specific transaction by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.TransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_GantangSatria_MyBank-BE_pkg_dto_request.CreateTransactionRequest": {
            "type": "object",
            "required": [
                "account_id",
                "amount",
                "pin",
                "type"
            ],
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string",
                    "maxLength": 255
                },
                "destination_account_number": {
                    "type": "string"
                },
                "destination_bank_code": {
                    "type": "string"
                },
                "destination_name": {
                    "type": "string"
                },
                "merchant_category": {
                    "type": "string",
                    "maxLength": 100
                },
                "merchant_location": {
                    "type": "string",
                    "maxLength": 200
                },
                "merchant_name": {
                    "type": "string",
                    "maxLength": 200
                },
                "note": {
                    "type": "string",
                    "maxLength": 255
                },
                "pin": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "TRANSFER",
                        "PAYMENT",
                        "TOPUP",
                        "WITHDRAW",
                        "DEPOSIT",
                        "QRIS"
                    ]
                }
            }
        },
...
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.Base": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.CategorySpendingResponse": {
            "type": "object",
            "properties": {
                "merchant_category": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "transaction_count": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.MerchantSummaryResponse": {
            "type": "object",
            "properties": {
                "merchant_category": {
                    "type": "string"
                },
                "merchant_name": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "transaction_count": {
                    "type": "integer"
                }
            }
        },
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.SpendingSummaryResponse": {
            "type": "object",
            "properties": {
                "category_breakdown": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.CategorySpendingResponse"
                    }
                },
                "fav_category": {
                    "type": "string"
                },
                "fav_method": {
                    "type": "string"
                },
                "top_merchants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.MerchantSummaryResponse"
                    }
                },
                "total_spend": {
                    "type": "number"
                },
                "total_transactions": {
                    "type": "integer"
                },
                "weekly_comparison": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GantangSatria_MyBank-BE_pkg_dto_response.WeeklySpendingResponse"
                    }
                }
            }
        },
...
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.TransactionResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "balance_after": {
                    "type": "number"
                },
                "balance_before": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "destination_account_number": {
                    "type": "string"
                },
                "destination_bank_code": {
                    "type": "string"
                },
                "destination_name": {
                    "type": "string"
                },
                "fee": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "is_recommended": {
                    "type": "boolean"
                },
                "merchant_category": {
                    "type": "string"
                },
                "merchant_location": {
                    "type": "string"
                },
                "merchant_name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "recommendation_id": {
                    "type": "integer"
                },
                "reference_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transacted_at": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_GantangSatria_MyBank-BE_pkg_dto_response.WeeklySpendingResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "current_week_spend": {
                    "type": "number"
                },
                "previous_week_spend": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}